+++ twitlib.tcl
@@ -21,6 +21,13 @@ namespace eval ::twitlib {
 	# Last tweet id we have seen - mentions timeline.
 	variable last_mentions_id 1
 
+	# Authenticated account user ID. This gets set automatically.
+	#
+	# TODO(horgh): We may need to re-set this back to 0 when we authenticate.
+	# e.g. if we switch between accounts when already running, we'll still have
+	# the first ID cached right now.
+	variable my_user_id 0
+
 	# OAuth authentication information.
 	variable oauth_consumer_key {}
 	variable oauth_consumer_secret {}
@@ -35,8 +42,9 @@ namespace eval ::twitlib {
 	# Create a tweet (new status).
 	variable status_url       https://api.twitter.com/2/tweets
 
-	# GET home_url to retrieve tweets by users you follow/yourself.
-	variable home_url         https://api.twitter.com/1.1/statuses/home_timeline.json
+	# Retrieve tweets by users you follow/yourself.
+	variable home_url         https://api.twitter.com/2/users/%d/timelines/reverse_chronological
+
 	variable mentions_url     https://api.twitter.com/1.1/statuses/mentions_timeline.json
 	variable msg_url          https://api.twitter.com/1.1/direct_messages/new.json
 	variable msgs_url         https://api.twitter.com/1.1/direct_messages.json
@@ -152,11 +160,12 @@ proc ::twitlib::get_my_screen_name {} {
 # get invalid ones.
 proc ::twitlib::fix_status {status} {
 	set changed 0
-	set tweet [dict get $status full_text]
+	set tweet [dict get $status text]
 
 	# if it has a retweet then as they can be truncated and lose
 	# data, especially urls, replace the tweet with the
 	# original tweet but add 'RT @name' to it.
+	# TODO(horgh): This is currently dead code for API v2.
 	if {[dict exists $status retweeted_status]} {
 		set rt_user [dict get $status retweeted_status user screen_name]
 		set rt_tweet [dict get $status retweeted_status full_text]
@@ -208,7 +217,7 @@ proc ::twitlib::fix_status {status} {
 	}
 
 	if {$changed} {
-		dict set status full_text $tweet
+		dict set status text $tweet
 	}
 	return $status
 }
@@ -251,27 +260,51 @@ proc ::twitlib::fix_statuses {statuses} {
 #     first request to get another 'page' of results.
 #     note I do not implement this here.
 proc ::twitlib::get_unseen_updates {} {
-	set params [list \
-		count $::twitlib::max_updates \
-		since_id $::twitlib::last_id \
-		tweet_mode extended \
+	if {$::twitlib::my_user_id == 0} {
+		set response [::twitlib::get_account_settings]
+		set ::twitlib::my_user_id [dict get $response body data id]
+	}
+
+	set url [format $::twitlib::home_url $::twitlib::my_user_id]
+	set body {}
+	set query_params [list \
+		max_results  $::twitlib::max_updates \
+		since_id     $::twitlib::last_id \
+		user.fields  id,username \
+		expansions   author_id \
+		tweet.fields author_id,created_at,text \
 	]
 
-	set result [::twitlib::query $::twitlib::home_url $params GET]
+	set result [::twitlib::query_v2 $url $body GET $query_params]
 
-	# re-order - oldest to newest.
-	set result [lreverse $result]
+	set status [dict get $result status]
+	set body [dict get $result body]
+	if {$status != 200} {
+		error "HTTP request failure: HTTP $status: $body"
+	}
+
+	set statuses [dict get $body data]
+	set includes [dict get $body includes]
 
 	# fix issues with truncation.
-	set statuses [::twitlib::fix_statuses $result]
+	#
+	# TODO(horgh): I don't know if this is necessary with API v2.
+	set statuses [::twitlib::fix_statuses $statuses]
 
 	set updates [list]
 	foreach status $statuses {
-		set user_id     [dict get $status user id_str]
-		set screen_name [dict get $status user screen_name]
-		set id          [dict get $status id_str]
+		set user_id     [dict get $status author_id]
+		set id          [dict get $status id]
 		set created_at  [dict get $status created_at]
-		set full_text   [dict get $status full_text]
+		set full_text   [dict get $status text]
+
+		set screen_name $user_id
+		foreach user [dict get $includes users] {
+			if {[dict get $user id] == $user_id} {
+				set screen_name [dict get $user username]
+				break
+			}
+		}
 
 		set d [dict create]
 		dict set d user_id     $user_id